// src/lib/types.ts
export type UserRole = 'admin' | 'editor' | 'viewer' | 'user';

export interface User {
  id: string;
  email: string;
  username: string; // Required in the interface
  name: string;
  role: UserRole;
  createdAt: string;
  displayName: string; // Required in the interface
  photoURL: string; // Required in the interface
}